[WRFE-70](feat): 상품 생성 로직 변경
