+++ apps/front/wraffle-webview/src/features/product/create/hooks/useCreateQuery.ts
@@ -1,25 +1,37 @@
-import type {
-  EventPayload,
-  Payload,
-  ProductResponse,
-  RafflePayload,
-} from '../config/type';
-import apiClient from '@/shared/api/apiClient';
+import {createProduct} from '../api/products';
+import {useRouter} from 'next/navigation';
 import {useMutation} from '@tanstack/react-query';
+import {useToast} from '@wraffle/ui';
 
-export const createProduct = async (payload: Payload) => {
-  const {type, ...data} = payload;
-  const url = type === 'raffle' ? '/raffles' : '/events';
+interface UseCreateProductProps {
+  type: 'event' | 'raffle';
+}
 
-  const response = await apiClient.post<
-    ProductResponse,
-    RafflePayload | EventPayload
-  >(url, {body: data, withAuth: true});
-  return response.data;
-};
-
-export const useCreateProductMutation = () => {
+export const useCreateProduct = ({type}: UseCreateProductProps) => {
+  const {toast} = useToast();
+  const router = useRouter();
   return useMutation({
-    mutationFn: (data: Payload) => createProduct(data),
+    mutationFn: createProduct,
+    onSuccess: response => {
+      toast({
+        title: '상품 생성 성공',
+        duration: 2000,
+        variant: 'success',
+        icon: 'check',
+      });
+      const thumbnailFileName = response.data.thumbnail.split('/').pop();
+      const productId = response.data.id;
+      router.push(
+        `/products/create/success?id=${productId}&thumbnail=${thumbnailFileName}&type=${type}`,
+      );
+    },
+    onError: error => {
+      toast({
+        title: (error as Error).message,
+        duration: 2000,
+        variant: 'warning',
+        icon: 'close',
+      });
+    },
   });
 };
